mtd: mtdoops: cast reason to int to fix enum comparison warnings

+++ drivers/mtd/mtdoops.c
@@ -431,12 +431,12 @@ static void mtdoops_do_dump(struct kmsg_dumper *dumper,
 	if(cxt->mtd == NULL)
 		return;
 
-	if(reason == KMSG_DUMP_SHUTDOWN ||
-		reason == KMSG_DUMP_EMERG)
+	if ((int)reason == (int)KMSG_DUMP_SHUTDOWN ||
+		(int)reason == (int)KMSG_DUMP_EMERG)
 		return;
 
 	/* Only dump oopses if dump_oops is set */
-	if (reason == KMSG_DUMP_OOPS && !dump_oops)
+	if ((int)reason == (int)KMSG_DUMP_OOPS && !dump_oops)
 		return;
 
 	do_dump_count++;
@@ -495,7 +495,7 @@ static void mtdoops_do_dump(struct kmsg_dumper *dumper,
 	} else
 		pr_err("mtdoops: read pmsg failed sig = 0x%x \n", p_hdr->sig);
 
-	if (reason == KMSG_DUMP_OOPS || reason == KMSG_DUMP_PANIC) {
+	if ((int)reason == (int)KMSG_DUMP_OOPS || (int)reason == (int)KMSG_DUMP_PANIC) {
 		/* Panics must be written immediately */
 		mtdoops_write(cxt, 1);
 	} else {
